Major changes

+++ MarketApplicationMVC.Application/ViewModel/Market/OfferDetailsVm.cs
@@ -13,7 +13,7 @@ namespace MarketApplicationMVC.Application.ViewModel.Market
         public string Name { get; set; }
         public byte[] Picture { get; set; }
         public string Description { get; set; }
-        public decimal Price { get; set; }
+        public int Price { get; set; } // rozwiązać problem z przecinkiem
         public string OfferCategory { get; set; }
         public int SellerName { get; set; }
